Monkeypatch ActionController::TestCase after app has loaded
Otherwise, ApplicationController::TestCase gets loaded before
the application is initialized and this results in weird
behavior in some tests, particularly view specs.
