+++ lib/vanity/frameworks/rails.rb
@@ -304,22 +304,6 @@ if defined?(ActionMailer)
   end
 end
 
-# Automatically configure Vanity.
-if defined?(Rails)
-  if Rails.const_defined?(:Railtie) # Rails 3
-    class Plugin < Rails::Railtie # :nodoc:
-      initializer "vanity.require" do |app|
-        Vanity::Rails.load!
-      end
-    end
-  else
-    Rails.configuration.after_initialize do
-      Vanity::Rails.load!
-    end
-  end
-end
-
-
 # Reconnect whenever we fork under Passenger.
 if defined?(PhusionPassenger)
   PhusionPassenger.on_event(:starting_worker_process) do |forked|
